pyproject.toml
[tool.poetry]
name = "alpha-mini"
version = "0.1.0"
description = "GulfCast: subseasonal risk pipeline."
authors = ["Xabier Blazquez <xabierblazq@gmail.com>"]
readme = "README.md"
packages = [{ include = "services" }]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26"
pandas = "^2.2"
xarray = "^2024.3"
zarr = "^2.16"
netcdf4 = "^1.6"
scipy = "^1.13"
typer = "^0.12"
requests = "^2.31"
httpx = "^0.27"
aiohttp = "^3.9"
fsspec = "^2024.3"
s3fs = ">=2025.7.0"   # bump from ^2024.3
geopandas = "^0.14"
pyarrow = "^16.0"
duckdb = "^0.10"
scikit-learn = "^1.4"
lightgbm = "^4.3"
quantile-forest = "^1.2"
joblib = "^1.4"
fastapi = "^0.110"
uvicorn = "^0.29"
jinja2 = "^3.1"
rich = "^13.7"
matplotlib = "^3.8"
plotly = "^5.20"
seaborn = "^0.13"
pyyaml = "^6.0"
pydantic = "^2.6"
cdsapi = ">=0.7,<0.8"
cfgrib = "^0.9.11"
torch = { version = "^2.2", optional = true }
torchvision = { version = "^0.17", optional = true }
openai = "^1.54"
physrisk-lib = { git = "https://github.com/os-climate/physrisk.git", rev = "91852f55c288319b00930c9607e0953afc050d76", optional = true }
mlflow = "^3.6.0"

[tool.poetry.extras]
gpu = ["torch", "torchvision"]
risk = ["xclim", "physrisk-lib"]

[tool.poetry.group.dev.dependencies]
black = "^24.4"
ruff = "^0.4"
ipykernel = "^6.29"
jupyter = "^1.0"
mypy = "^1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
